requirements.txt
# LLM & Agent
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Data
pandas>=2.2.0
openpyxl>=3.1.0
numpy>=1.26.0

# Charts
matplotlib>=3.8.0
plotly>=5.22.0
kaleido>=0.2.1          # export plotly a immagine

# Telegram
python-telegram-bot>=21.0

# Utils
python-dotenv>=1.0.0
requests>=2.32.0
rich>=13.7.0             # output CLI colorato

# (Futuro - FastAPI)
# fastapi>=0.115.0
# uvicorn>=0.30.0
